requirements.txt
mpmath==1.3.0
PyYAML==6.0.3
sympy==1.14.0
